requirements.txt
fastapi 
uvicorn 
sqlalchemy
pymysql 
python-dotenv 
pyjwt
python-multipart
passlib
passlib[bcrypt]